+++ src/strategies/jpegd-locked-jpeg-of/index.ts
